refactor: live switch button
